Move SetupFreeListings component of editing free listings to the shared directory
