+++ js/src/edit-free-campaign/index.js
@@ -14,10 +14,10 @@ import { isEqual } from 'lodash';
 import { useAppDispatch } from '.~/data';
 import TopBar from '.~/components/stepper/top-bar';
 import ChooseAudience from '.~/components/free-listings/choose-audience';
+import SetupFreeListings from '.~/components/free-listings/setup-free-listings';
 import useTargetAudienceFinalCountryCodes from '.~/hooks/useTargetAudienceFinalCountryCodes';
 import useSettings from '.~/components/free-listings/configure-product-listings/useSettings';
 import useApiFetchCallback from '.~/hooks/useApiFetchCallback';
-import SetupFreeListings from './setup-free-listings';
 import useLayout from '.~/hooks/useLayout';
 import useNavigateAwayPromptEffect from '.~/hooks/useNavigateAwayPromptEffect';
 import useShippingRates from '.~/hooks/useShippingRates';
